Mark MiddlewareRunner as internal and move to Io namespace
