+++ tests/benchmark-middleware-runner.php
@@ -1,8 +1,8 @@
 <?php
 
 use Psr\Http\Message\ServerRequestInterface;
+use React\Http\Io\MiddlewareRunner;
 use React\Http\Io\ServerRequest;
-use React\Http\MiddlewareRunner;
 use React\Http\Response;
 
 const ITERATIONS = 5000;
